backend-py/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
groq>=0.11.0
pinecone>=5.0.0
python-dotenv>=1.0.1
httpx>=0.27.2
pydantic>=2.9.2
deepeval>=1.5.2
langfuse>=2.53.0
openai>=1.51.0
sentence-transformers>=3.0.0
